# contracts/openapi.json
{
  "components": {
    "schemas": {
      "AnswerIn": {
        "additionalProperties": false,
        "properties": {
          "is_correct": {
            "title": "Is Correct",
            "type": "boolean"
          },
          "score_weight": {
            "description": "Points awarded when matched.",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score Weight",
            "type": "integer"
          },
          "text": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "is_correct",
          "score_weight"
        ],
        "title": "AnswerIn",
        "type": "object"
      },
      "CaseCreated": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "CaseCreated",
        "type": "object"
      },
      "CaseList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CaseSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "limit",
          "offset"
        ],
        "title": "CaseList",
        "type": "object"
      },
      "CaseStats": {
        "description": "Aggregate attempt data. Carries no answers, so it cannot leak the key.",
        "properties": {
          "average_score": {
            "title": "Average Score",
            "type": "number"
          },
          "max_score": {
            "title": "Max Score",
            "type": "integer"
          },
          "outcomes": {
            "items": {
              "$ref": "#/components/schemas/OutcomeCount"
            },
            "title": "Outcomes",
            "type": "array"
          },
          "submissions": {
            "title": "Submissions",
            "type": "integer"
          }
        },
        "required": [
          "submissions",
          "outcomes",
          "average_score",
          "max_score"
        ],
        "title": "CaseStats",
        "type": "object"
      },
      "CaseSummary": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "created_at"
        ],
        "title": "CaseSummary",
        "type": "object"
      },
      "ClinicalCaseCreate": {
        "additionalProperties": false,
        "properties": {
          "answers": {
            "description": "Answer key. Stored server-side only; never returned by public endpoints.",
            "items": {
              "$ref": "#/components/schemas/AnswerIn"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Answers",
            "type": "array"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/FindingIn"
            },
            "maxItems": 50,
            "minItems": 1,
            "title": "Findings",
            "type": "array"
          },
          "patient_age": {
            "anyOf": [
              {
                "maximum": 130.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Age"
          },
          "patient_sex": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PatientSex"
              },
              {
                "type": "null"
              }
            ]
          },
          "presentation": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Presentation",
            "type": "string"
          },
          "provenance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Provenance"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "presentation",
          "findings",
          "answers"
        ],
        "title": "ClinicalCaseCreate",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "FindingCategory": {
        "enum": [
          "history",
          "symptom",
          "vital_sign",
          "physical_exam",
          "laboratory",
          "imaging",
          "other"
        ],
        "title": "FindingCategory",
        "type": "string"
      },
      "FindingIn": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "$ref": "#/components/schemas/FindingCategory"
          },
          "value": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "category",
          "value"
        ],
        "title": "FindingIn",
        "type": "object"
      },
      "FindingMeasurement": {
        "description": "Numeric value parsed from the finding text, with its reference range.\n\nDerived and optional: the finding text remains the source of truth. This\nexists so the interface can plot a value instead of burying it in a\nsentence. Reference ranges are adult values shown as teaching context.",
        "properties": {
          "axis_max": {
            "title": "Axis Max",
            "type": "number"
          },
          "axis_min": {
            "title": "Axis Min",
            "type": "number"
          },
          "flag": {
            "$ref": "#/components/schemas/MeasurementFlag"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "reference_high": {
            "title": "Reference High",
            "type": "number"
          },
          "reference_low": {
            "title": "Reference Low",
            "type": "number"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "label",
          "value",
          "unit",
          "reference_low",
          "reference_high",
          "axis_min",
          "axis_max",
          "flag"
        ],
        "title": "FindingMeasurement",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HealthStatus": {
        "properties": {
          "status": {
            "enum": [
              "ok",
              "unavailable"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "HealthStatus",
        "type": "object"
      },
      "MeasurementFlag": {
        "enum": [
          "low",
          "normal",
          "high"
        ],
        "title": "MeasurementFlag",
        "type": "string"
      },
      "OutcomeCount": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "outcome": {
            "$ref": "#/components/schemas/ScoreOutcome"
          }
        },
        "required": [
          "outcome",
          "count"
        ],
        "title": "OutcomeCount",
        "type": "object"
      },
      "PatientSex": {
        "enum": [
          "female",
          "male",
          "other"
        ],
        "title": "PatientSex",
        "type": "string"
      },
      "Provenance": {
        "additionalProperties": false,
        "properties": {
          "model": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "prompt_version": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Version"
          },
          "source": {
            "enum": [
              "manual",
              "llm_extraction"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "source"
        ],
        "title": "Provenance",
        "type": "object"
      },
      "PublicClinicalCase": {
        "description": "What a participant sees. Deliberately has NO answer / correctness / weight fields.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/PublicFinding"
            },
            "title": "Findings",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "patient_age": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Age"
          },
          "patient_sex": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PatientSex"
              },
              {
                "type": "null"
              }
            ]
          },
          "presentation": {
            "title": "Presentation",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "patient_age",
          "patient_sex",
          "presentation",
          "findings",
          "created_at"
        ],
        "title": "PublicClinicalCase",
        "type": "object"
      },
      "PublicFinding": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/FindingCategory"
          },
          "measurements": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/FindingMeasurement"
            },
            "title": "Measurements",
            "type": "array"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "category",
          "value"
        ],
        "title": "PublicFinding",
        "type": "object"
      },
      "ScoreOutcome": {
        "enum": [
          "correct",
          "partially_correct",
          "incorrect"
        ],
        "title": "ScoreOutcome",
        "type": "string"
      },
      "ScoreRequest": {
        "additionalProperties": false,
        "properties": {
          "answer": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Answer",
            "type": "string"
          }
        },
        "required": [
          "answer"
        ],
        "title": "ScoreRequest",
        "type": "object"
      },
      "ScoreResponse": {
        "properties": {
          "feedback": {
            "title": "Feedback",
            "type": "string"
          },
          "is_correct": {
            "title": "Is Correct",
            "type": "boolean"
          },
          "max_score": {
            "title": "Max Score",
            "type": "integer"
          },
          "outcome": {
            "$ref": "#/components/schemas/ScoreOutcome"
          },
          "score": {
            "title": "Score",
            "type": "integer"
          },
          "submission_id": {
            "format": "uuid",
            "title": "Submission Id",
            "type": "string"
          }
        },
        "required": [
          "submission_id",
          "score",
          "max_score",
          "is_correct",
          "outcome",
          "feedback"
        ],
        "title": "ScoreResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-Admin-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "summary": "Clinical cases, answer keys and deterministic diagnosis scoring.",
    "title": "Eximion Clinical Case API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/cases": {
      "get": {
        "operationId": "list_cases_api_v1_cases_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 10000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List cases (newest first)",
        "tags": [
          "cases"
        ]
      },
      "post": {
        "operationId": "create_case_api_v1_cases_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClinicalCaseCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Create a clinical case with its answer key (authoring)",
        "tags": [
          "cases"
        ]
      }
    },
    "/api/v1/cases/{case_id}": {
      "get": {
        "operationId": "get_case_api_v1_cases__case_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicClinicalCase"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get a case for solving (never includes the answer key)",
        "tags": [
          "cases"
        ]
      }
    },
    "/api/v1/cases/{case_id}/score": {
      "post": {
        "operationId": "score_case_api_v1_cases__case_id__score_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit a diagnosis and receive a deterministic score",
        "tags": [
          "cases"
        ]
      }
    },
    "/api/v1/cases/{case_id}/stats": {
      "get": {
        "operationId": "case_stats_api_v1_cases__case_id__stats_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseStats"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Aggregate attempt statistics (counts only, never answers)",
        "tags": [
          "cases"
        ]
      }
    },
    "/health": {
      "get": {
        "operationId": "liveness_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthStatus"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Liveness: process is up (no dependencies checked)",
        "tags": [
          "health"
        ]
      }
    },
    "/health/ready": {
      "get": {
        "operationId": "readiness_health_ready_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthStatus"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Readiness: database reachable",
        "tags": [
          "health"
        ]
      }
    }
  }
}